+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## [Unreleased]
-* Bugfix: missing import class quickfix (#1132) 
+* Bugfix: "import class quickfix" missing in some cases (#1132) 
 * Bugfix: "Find Usages" was not checking catch blocks correctly(#929)
+* Bugfix: "Introduce Constant"  placed outside of class
+* Bugfix: "Pull member up" placed code outside of class
 * Change: hiding haxe context menus in non-haxe projects
 
 ## 1.4.11